# doc/api.yaml
openapi: 3.0.0
info:
  title: WASATEXT
  description: API for WASATEXT messaging platform
  version: 1.0.0
...
security:
  - WASATextAuth: []
paths:
...
  /users/{userId}/conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getMessages
      tags:
        - "Conversations"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                description: List of messages
                properties:
                  messages:
                    description: Message schema
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Message"
                        - properties:
                            sender:
                              allOf:
                                - $ref: "#/components/schemas/User"
                                - description: Sender of the message, you or other.
              example:
                messages:
                  - body: "Hello!"
                    messageId: 1
                    time: "2025-10-31T15:42:00Z"
                    read: "received"
                    sender:
                      userId: 11300
                      username: Simone

        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Send new direct message
      description: Send new message in the conversation.
      operationId: sendMessage
      tags:
        - "Message"

      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Send"
        required: true
      responses:
        "201":
          description: Message sent. By default, the message has "read" set to sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Message"
                  - type: object
                    properties:
                      read:
                        description: Show the message status
                        type: string
                        minLength: 0
                        maxLength: 255
                        pattern: "^[a-zA-Z0-9_ -]+$"
                        enum: [ received ]
                        example: received
              example:
                body: "Hi!"
                messageId: 4444
                time: "2025-10-31T15:42:00Z"
                read: "received"

        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /users/{userId}/groups/{groupId}/info/name:
    put:
      summary: Set Group name
      description: The user can set a personalized group name.
      operationId: setGroupName
      tags:
        - "Group"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        content:
          application/json:
            schema:
              description: Schema for updating the group's name
              properties:
                name:
                  description: Group's name
                  type: string
                  minLength: 1
                  maxLength: 255
                  pattern: "^[a-zA-Z0-9_ -]+$"
                  example: capodanno 2027
      responses:
        "201":
          description: Name updated.
          content:
            application/json:
              schema:
                description: Updated group name response schema
                properties:
                  name:
                    description: Group's name
                    type: string
                    minLength: 1
                    maxLength: 255
                    pattern: "^[a-zA-Z0-9_ -]+$"
                    example: capodanno_2027
                  time:
                    description: Timestamp
                    type: string
                    format: date-time
                    example: 2025-10-31T15:42:00Z
                    minLength: 20
                    maxLength: 20
                    pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"
              example:
                name: "capodanno_2027"
                time: "2025-10-31T15:42:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups/{groupId}:
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroup
      tags:
        - "Conversations"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: 288344
                name: "capodanno2026"
                upload:
                  url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                  mime: "image/jpg"
                  width: 1024
                  height: 720
                participants:
                  - userId: 10012
                    username: "Simone"
                  - userId: 77772
                    username: "Alessio"


        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /users/{userId}/groups/{groupId}/messages:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/GroupId"

    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroupMessages
      tags:
        - "Conversations"

      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                description: Response schema. returns che chat
                properties:
                  messages:
                    description: List of messages
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                      allOf:
                        - $ref: "#/components/schemas/Message"
                        - properties:
                            sender:
                              allOf:
                                - $ref: "#/components/schemas/User"
                                - description: Sender of the message, you or other.
              example:
                messages:
                  - body: "Benvenuti nel gruppo!"
                    messageId: 123456
                    time: "2025-10-31T15:42:00Z"
                    read: "received"
                    sender:
                      userId: 11300
                      username: "Simone"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Send new group message
      description: Send new message in the group.
      operationId: sendGroupMessage
      tags:
        - "Message"


      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Send"

        required: true

      responses:
        "201":
          description: Message sent. By default, the message has "read" set to sent.The message has "read" when all members of group has read the message.
          content:
            application/json:
              schema:
                description: Response schema
                allOf:
                  - $ref: "#/components/schemas/Message"
                  - properties:
                      read:
                        description: represents the status of the message
                        type: string
                        minLength: 0
                        maxLength: 255
                        pattern: "^[a-zA-Z0-9_ -]+$"
                        enum: [ received ]
                        example: received
              example:
                body: Hellooooo
                messageId: 4333744
                time : 2025-10-31T15:42:00Z
                read: received


        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  schemas:
...
    UserId:
      description: Operation User's ID
      type: integer
      example: 11300
      readOnly: true
    Username:
      description: User's username
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[a-zA-Z0-9_ -]+$"
      example: Simone
    User:
      description: User schema
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userId
        - username
    Group:
      type: object
      description: group schema
      properties:
        groupId:
          $ref: "#/components/schemas/GroupId"
        name:
          description: Group name
          type: string
          minLength: 0
          maxLength: 255
          pattern: "^[a-zA-Z0-9_ -]+$"
          example: capodanno2026
        upload:
          $ref: "#/components/schemas/Upload"
        participants:
          description: list of participants of a group
          type: array
          minItems: 2
          maxItems: 256
          items:
            $ref: "#/components/schemas/User"
          example:
            - userId: 11300
              username: Simone
            - userId: 11301
              username: Luca
    ConversationId:
      description: Unique ID assigned to each conversation resource in the system.
      type: integer
      example: 110056
      readOnly: true
...
    MessageId:
      description: Unique ID assigned to each message resource in the system.
      type: integer
      example: 2635444
      readOnly: true
    GroupId:
      description: Unique ID assigned to each group resource in the system.
      type: integer
      example: 2773354
      readOnly: true
...
    Photo:
      description: Binary image file
      type: string
      format: binary
      pattern: "^[A-Za-z0-9+/=]+$"
      minLength: 1
      maxLength: 9999999999999999999
    Upload:
      description: upload's schema for responses or view
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-zA-Z]+://.+$"
          format: url
          description: Photo's URL to access the image
          example: https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg
        mime:
          description: mime represents the type of the Photo(image/png...)
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^[a-zA-Z]+/[a-zA-Z.]+$'
          example: image/jpg
        width:
          description: width of the image
          type: integer
          example: 1024
        height:
          description: height of the image
          type: integer
          example: 720
    Send:
      type: object
      description: What to send, Image or text
      properties:
        body:
          oneOf:
            - $ref: "#/components/schemas/Text"
            - $ref: "#/components/schemas/PhotoMessage"
    Message:
      type: object
      allOf:
        - $ref: "#/components/schemas/Send"
        - properties:
            messageId:
              $ref: "#/components/schemas/MessageId"
            time:
              description: Timestamp
              type: string
              format: date-time
              example: 2025-10-31T15:42:00Z
              minLength: 20
              maxLength: 20
              pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$"

            read:
              type: string
              minLength: 0
              maxLength: 255
              description: Read state of the message
              pattern: "^[a-zA-Z0-9_ -]+$"
              enum: [ received, read ]
              example: received
...
    Text:
      description: Text of the message
      type: string
      minLength: 0
      maxLength: 2048
      example: Hello, World!
      pattern: "^[a-zA-Z0-9 .,!?_-]+$"
...
    PhotoMessage:
      type: object
      description: The user can send an image with an optional description.
      allOf:
        - $ref: "#/components/schemas/Photo"
        - properties:
            caption:
              $ref: "#/components/schemas/Text"
    Error:
      description: Info of an error
      type: object
      properties:
        code:
          description: code of the error
          type: integer
          example: 401
        message:
          description: text of the error
          type: string
          minLength: 0
          maxLength: 2048
          example: Unauthorized Action, request invalid or missing.
      required:
        - code
        - message
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User Identifier
      schema:
        $ref: "#/components/schemas/UserId"
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Chat Identifier
      schema:
        $ref: "#/components/schemas/ConversationId"
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group Identifier
      schema:
        $ref: "#/components/schemas/GroupId"
  securitySchemes:
    WASATextAuth:
      type: http
      description: Protect operation session token.
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    UnauthorizedError:
      description: Unauthorized Action, request invalid or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: "Unauthorized Action, request invalid or missing."
    InternalError:
      description: An error occurred, please retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: "An error occurred, please retry."
    NotFoundError:
      description: Resource not found. Check the parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: "Resource not found. Check the parameters."
    BadRequest:
      description: Bad Request. Invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 400
            message: "Bad Request. Invalid parameters."
...